F: DB: can add sqrl identities after the fact

+++ java/qrauth-server/src/main/java/com/allogy/qrauth/server/entities/AuthMethod.java
@@ -177,9 +177,9 @@ enum AuthMethod
 	}
 
 	public
-	long getDefaultLoginLength()
+	int getDefaultLoginLength()
 	{
-		return FIRST_RANK_MILLIS/rank;
+		return (int)(FIRST_RANK_MILLIS/rank);
 	}
 
 	/**
